split api and slices
