+++ src/hooks/useAuth.tsx
@@ -1,4 +1,4 @@
-import { useGetUserDataQuery } from '@/redux/slices/user-slice';
+import { useGetUserDataQuery } from '@/redux/api/user-api';
 
 export const useAuth = () => {
   const { data = null } = useGetUserDataQuery('');
