+++ src/redux/index.ts
@@ -2,9 +2,9 @@ import { configureStore } from '@reduxjs/toolkit';
 import { fieldOnFocusReducer } from './slices/focusedFields-slice';
 import { modalReducer } from './slices/modal-slice';
 import current from './slices/current-slice';
-import { diziIziSplitApi } from './slices/diziIziSplitApi-slice';
-import { CitiesApi } from './slices/cities-slice';
-import { yandexAuthApi } from './slices/yandexAuthApi';
+import { diziIziSplitApi } from './api/diziIzi-splitApi';
+import { CitiesApi } from './api/cities-api';
+import { yandexAuthApi } from './api/yandexAuth-api';
 
 export const store = configureStore({
   reducer: {
